# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"GetEmailResponseData": {
				"properties": {
					"email": {
						"type": "string",
						"description": "The user's email",
						"example": "test@clippic.app",
						"maxLength": 200
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetEmailResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetEmailResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutEmailResponseData": {
				"properties": {
					"email": {
						"type": "string",
						"description": "The user's email",
						"example": "test@clippic.app",
						"maxLength": 200
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutEmailResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PutEmailResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ClippicErrorInterface": {
				"properties": {
					"status": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32"
					},
					"trace": {
						"type": "string"
					}
				},
				"required": [
					"message",
					"code",
					"trace"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MailFormatError": {
				"$ref": "#/components/schemas/ClippicErrorInterface",
				"description": "1022 | Error when access mode is not valid."
			},
			"PutEmailRequest": {
				"description": "Update the users email",
				"properties": {
					"email": {
						"type": "string",
						"description": "The user's email",
						"example": "testnew@clippic.app",
						"minLength": 5,
						"maxLength": 200
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"email": "testnew@clippic.app"
				}
			},
			"GetForenameResponseData": {
				"properties": {
					"forename": {
						"type": "string",
						"description": "The user's forename",
						"example": "tester",
						"maxLength": 100
					}
				},
				"required": [
					"forename"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetForenameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetForenameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutForenameResponseData": {
				"properties": {
					"forename": {
						"type": "string",
						"description": "The user's forename",
						"example": "tester",
						"maxLength": 100
					}
				},
				"required": [
					"forename"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutForenameResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PutForenameResponseData"
						},
						"type": "array"
					},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutForenameRequest": {
				"description": "Create or update the users forename",
				"properties": {
					"forename": {
						"type": "string",
						"description": "The user's forename",
						"example": "tester",
						"minLength": 1,
						"maxLength": 100
					}
				},
				"required": [
					"forename"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"username": "tester"
				}
			},
...
			"ClippicResponse": {
				"properties": {
					"status": {
						"type": "string",
						"description": "Respose status, either \"error\" or \"success\""
					},
					"message": {
						"type": "string",
						"description": "Custom response message."
					},
					"data": {},
					"code": {
						"type": "integer",
						"format": "int32",
						"description": "HTTP code or custom error code."
					},
					"trace": {
						"type": "string",
						"description": "OpenTracing tracing id for request."
					}
				},
				"required": [
					"status",
					"code",
					"trace"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PasswordInvalidError": {
				"$ref": "#/components/schemas/ClippicErrorInterface",
				"description": "1033 | Error when password is not passing security RegEx."
			},
			"PutPasswordRequest": {
				"description": "Update the users password.",
				"properties": {
					"pass": {
						"type": "string",
						"description": "The user's password.",
						"example": "Test1234#!"
					}
				},
				"required": [
					"pass"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"pass": "Test1234#!"
				}
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "@clippic/user-v2",
		"version": "0.0.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users/v2/email": {
			"get": {
				"operationId": "GetEmailRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetEmailResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"email": "test@clippic.app"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will return the user's Email.",
				"tags": [
					"Email"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutEmailRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PutEmailResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"email": "testnew@clippic.app"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MailFormatError"
								}
							}
						}
					}
				},
				"description": "This request will update the user's email.\n\n**Errors:**\n\n| Code | Description                 |\n|------|-----------------------------|\n| 400  | AccountAlreadyExistsError   |\n| 400  | MailFormatError             |",
				"tags": [
					"Email"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutEmailRequest"
							}
						}
					}
				}
			}
		},
		"/users/v2/forename": {
			"get": {
				"operationId": "GetForenameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetForenameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"forename": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will return the user's forename.",
				"tags": [
					"Forename"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "PutForenameRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PutForenameResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "",
											"data": [
												{
													"forename": "tester"
												}
											],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will update or create if not exists the user's forename.",
				"tags": [
					"Forename"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutForenameRequest"
							}
						}
					}
				}
			}
		},
...
		"/users/v2/password": {
			"put": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Password changed successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClippicResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "Password changed successfully",
											"data": [],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PasswordInvalidError"
								}
							}
						}
					}
				},
				"description": "This request will update users password.\n\n**Errors:**\n\n| Code | Description                 |\n|------|-----------------------------|\n| 400  | PasswordInvalidError        |",
				"tags": [
					"Password"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "header",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutPasswordRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "ForgotPassword",
				"responses": {
					"200": {
						"description": "Password reset instructions has been sent if email address is registered.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ClippicResponse"
								},
								"examples": {
									"Example 1": {
										"value": {
											"status": "success",
											"message": "Password reset instructions has been sent if email address is registered.",
											"data": [],
											"code": 200,
											"trace": "4ba373202a8e4807"
										}
									}
								}
							}
						}
					}
				},
				"description": "This request will reset passwort after forgotten.",
				"tags": [
					"Password"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}